Songu1

+++ Backend/invitation/src/main/java/com/coho/invitation/controller/CardController.java
@@ -10,15 +10,16 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.userdetails.User;
 import org.springframework.web.bind.annotation.*;
 
-import java.io.IOException;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.UUID;
 
 @Tag(name="3) Card API",description = "초대장(청첩장,초대장) API")
-//@UserAuthorize
+@UserAuthorize
 @RestController
 @RequestMapping("/api/cards")
 @CrossOrigin(origins = "*")
@@ -36,8 +37,13 @@ public class CardController {
     @Operation(summary = "특정 행사의 초대장(카드) 조회", description = "파라미터로 받은 event id의 초대장 카드 정보를 반환")
     @Parameter(name="str", description = "선택한 행사의 event id를 전송")
     @GetMapping("/{eid}")
-    public ResponseEntity<?> getCard(@PathVariable("eid") String eid){
+    public ResponseEntity<?> getCard(@AuthenticationPrincipal User user, @PathVariable("eid") String eid){
         Map<String,Object> map = new HashMap<>();
+        String uid = user.getUsername();
+
+        // 현재 사용자가 행사 권한자 인지 확인
+        if(!eventService.checkAuthority(eid).contains(uid))
+            return ResponseEntity.ok().body("권한 없음");
 
         Event event = eventService.getEvent(eid);
         Card card = cardService.getCard(eid);
@@ -74,12 +80,18 @@ public class CardController {
     @Operation(summary = "특정 행사의 초대장 카드 추가", description = "파라미터로 받은 event id의 카드 정보를 사용하여 초대장을 저장하고 정보 반환")
     @Parameter(name="str", description = "선택한 행사의 event id를 전송")
     @PostMapping("/templates/{eid}")
-    public ResponseEntity<?> addCard(@PathVariable("eid") String eid, @RequestBody JsonNode params){
+    public ResponseEntity<?> addCard(@AuthenticationPrincipal User user, @PathVariable("eid") String eid, @RequestBody JsonNode params){
         Card card = new Card();
+
+        String uid = user.getUsername();
+
+        // 현재 사용자가 행사 권한자 인지 확인
+        if(!eventService.checkAuthority(eid).contains(uid))
+            return ResponseEntity.ok().body("권한 없음");
 
         /* 행사에 초대장 존재 여부 확인 */
         if(cardService.checkDuplicateCard(eid).isPresent())
-            return ResponseEntity.ok().body("이미 초대장이 존재합니다.");
+            return ResponseEntity.ok().body("초대장이 없습니다");
 
         /* 초대장 템플릿 생성 */
         card.setEid(eid);
@@ -98,15 +110,21 @@ public class CardController {
 
         cardService.addCardTemplate(card);
 
-        return ResponseEntity.ok().body(card.getCid());
+        return ResponseEntity.ok().body(card);
     }
 
     /* 초대장 수정하기 */
     @Operation(summary = "특정 초대장 카드의 정보 수정", description = "파라미터로 받은 card id와 초대장 카드 정보를 통해 수정하고 card id 반환")
     @Parameter(name="str", description = "선택한 초대장의 card id를 전송")
     @PutMapping("/{cid}")
-    public ResponseEntity<?> updateCard(@PathVariable("cid") String cid, @RequestBody JsonNode params){
+    public ResponseEntity<?> updateCard(@AuthenticationPrincipal User user, @PathVariable("cid") String cid, @RequestBody JsonNode params){
         Card card = new Card();
+        String uid = user.getUsername();
+
+        /* 카드 삭제 권한 체크 */
+        if(!cardService.checkCardAuthority(cid).contains(uid)){
+            return ResponseEntity.ok().body("권한 없음");
+        }
 
         card.setCid(cid);
         if(params.get("ctype").toString().equals("Template"))
@@ -128,7 +146,13 @@ public class CardController {
     @Operation(summary = "특정 초대장 카드 삭제", description = "파라미터로 받은 card id의 초대장 카드 삭제 후 card id 반환")
     @Parameter(name="str", description = "선택한 행사의 event id를 전송")
     @DeleteMapping("/{cid}")
-    public ResponseEntity<?> deleteCard(@PathVariable("cid") String cid){
+    public ResponseEntity<?> deleteCard(@AuthenticationPrincipal User user, @PathVariable("cid") String cid){
+        String uid = user.getUsername();
+
+        /* 카드 삭제 권한 체크 */
+        if(!cardService.checkCardAuthority(cid).contains(uid)){
+            return ResponseEntity.ok().body("권한 없음");
+        }
 
         cardService.deleteCard(cid);
